Added the ability to replicate over a socket.
This patch adds a new argument to the CHANGE MASTER syntax,
MASTER_SOCKET="/path/to/socket". If the hostname is "localhost" the
server will now connect and communicate with the master server through
the given socket. This also adds a new field in the master.info file
'socket=/path/to/socket' which preserves the master socket through resets.

Note that if the socket name is longer than the maximum allowed by the
addrinfo_un struct then the server will be unable to connect to the
master and will print an appropriate error message.

All logging messages related to replication have been modified to
show only the target they are actually replicating over, either
'user@host:port' for tcp or 'user@socket[/path/to/socket]' for socket.

Change-Id: I75e93e162737983aebcee1017abf4dd7ff0a2208