Added status variables for threads killed and runs completed.
Made status variables for the number of sniping runs completed
(Sniper_runs), the number of threads that the sniper has attempted
to kill (Sniper_queries_killed) and the number of threads each
type of sniper has approved for killing and were not vetoed on
(Sniper_{idle,connectionless,long,infeasible}_queries_killed). Updated
the tests to check these variables as well.

Change-Id: I3509f66fd133527a9f23c571be52214a56af56e1