Add --rpl-semi-sync-slave-reset-packet-num flag.
This flag will control whether semi-sync slave resets mysql protocol packets
numbering on each semi-sync ack. The reset was initially intended to be done
once for each command, so resetting on each semi-sync ack is kind of against
the protocol ideology. And also it's a crucial differenece between MariaDB's and
Google-MySQL's implementations of semi-sync replication.

This patch is upstreamable if they will be willing to accept such change
(the "bad" behavior exists at least since MySQL 5.5, commit in bzr is
from 2009). If they won't accept the change we'll set the flag to 0 during
migration, then set it to 1 and then get rid of it completely.

Change-Id: If27e73f6f291edca41f7459a6cda359a85f9fbca