Allow --replicate-same-sever-id to be used together with --log-slave-updates.
This is a temporary patch to allow these two flags to be used together until
upstream figures out how to deal with the problem properly. The problem is
when some server is master, logs some events, then failover happens to another
server and then first server is restored from backup and should replay some
events that it logged when it was a master. Without --replicate-same-server-id
server will skip all events it created when it was a master.

Change-Id: I2184aba0d702a51ece2c48867431385ee740370a