Make join_outer_jcl6 stable by adding extra column to order by
join_outer_jcl6 performed sorting on one column with multiple rows
having same value. For some cases this was being run on a heap table
with filesort using references (row pointers). For heap tables, the
reference is an actual memory pointer, which made the sorting
non deterministic. Solution is to add an extra column to the
order by which does not modify the execution plan.

Change-Id: Ic4d57184eb079d6eb1c360550fb03d7d245316fd
Reviewed-by: Pavel Ivanov <pivanof@google.com>