Fix compile error in slave.cc
In slave.cc code is added to handle the new event types added
in MariaDB. These even types have numbers >= 160. However,
in slave.cc a switch was made using a signed char for event-type
and since signed char has a range of -128 to 127 these are not
in range. This resulted in compile error.

Change-Id: Ifc1dad34161977fcf5c11f27f7dae4c789565d96
Reviewed-by: Pavel Ivanov <pivanof@google.com>