Fix user and hostname used by automatic_sp_privileges
When a user without global EXECUTE privileges creates a stored procedure
with automatic_sp_privileges mode enabled, a procedure-specific EXECUTE
privilege would be automatically granted on that user. In the function
which handles this automatic grant, the user and host used for the grant
were derived from the actual connected user and resolved hostname, not
on the stored ACL under which the user connected.

If a user was connected using a "fuzzy" ACL, such as a host containing a
wildcard, this could cause a completely new user (with the exact host of
the connected user) to be auto-created. In practice this would mean the
automatic privileges granted could actually cause the user in question
to *lose* privileges (having only USAGE) when connecting from the same
specific host, on the global context, while gaining the intended
procedure-specific EXECUTE privilege.

The correct behavior is always to add the automatic grant on the user
as-authenticated, which is stored in the Security_context as priv_user
and priv_host.

Additionally, a test for MySQL Bug 44658 was removed from grant.test
because it was testing behavior that only existed due to this bug.

Change-Id: I14a32a4392d4e1a8bad38276f17a071280fc1b09
Reviewed-by: Pavel Ivanov <pivanof@google.com>