Port semi-sync fixes from MySQL 5.1.
- Handle KILL of waiting threads in semi-sync.

- Do not automatically turn off semi-sync master. Semi-sync replication on
master was automatically turned off every time master hit timeout waiting for
semi-sync ack or if thread waiting for semi-sync ack was killed. This change
doesn't ever turn off semi-sync replication if
semi_sync_master_wait_no_slaves = ON. This change also deletes tests that were
testing exactly this behavior -- automatic turning off of semi-sync replication
and returning "success" to client even when transaction wasn't semi-sync ack'ed.

- Harden semi-sync masters against bad semi-sync packets. Sanity check the log
file name and position in semi-sync slave replies so that bad data has less
chance of breaking the master's state.

- Return error to clients on semi-sync repl failures. This patch adds error
reporting of semi-sync failures, when the connection was killed before
the master could receive an ACK from one of the slaves. When a client receives
this error, the transaction has been committed to the master, but not
necessarily to any of the slaves. If this happens, it could mean that the master
was already replaced, and that this transaction is an alternate future. Manual
cleanup is required in this scenario. The aim of this patch is to propagate
the error up to the clients, in the hopes that it is noticed.

Also new code was added to make sure that when transaction didn't get
a semi-sync ack client connection gets killed. And flakiness of the test was
fixed.

Change-Id: I9827b87b95753ad35bd903f85430cc4dadf835df