Do not include the selected and updated column list into audit log.
During optimization stage of query execution MySQL can decide to use temporary
table for some of the results. And it will replace some of the column
references and expressions in the SELECT list with references to columns in the
temporary table. But by the time audit logging uses the items_list in the LEX
structure temporary tables are already destroyed and any references to columns
in the temporary table point to already freed memory. Thus audit logging
causes MySQL to crash in such cases.
Besides having a list of all selected columns and all expressions performed on
the selected columns doesn't have much of useful information for audit. So we'll
exclude it from now on.

Change-Id: I7222d7c31e807eb6ec5fd16cd3e209f955b969ac
Reviewed-by: Jeremy Cole <jeremycole@google.com>