Fix limit_tmp_disk_space.test on 32-bit systems
Set tmp_table_size to a low value to force the internal temporary
tables out to disk so that they are affected by limit_tmp_disk_space.
The heap-table used for memory version of internal temporary table
is slightly smaller on 32-bit, making testcase to fail as the
temporary table is never forced out to disk.

Change-Id: I90484669c8983140712dc218bf7dc5065a290418
Reviewed-by: Pavel Ivanov <pivanof@google.com>