Change the character set of the output of SHOW ENGINE [INNODB] STATUS
The "Status" field output by the SHOW ENGINE [INNODB] STATUS command was
being set to system_charset_info when it actually contains binary data
from other sessions and is in no way guaranteed to currently be or to
be convertible to utf8. It should be returned as my_charset_bin (binary)
instead.

Change-Id: I0f284156ff84404e1aaa4dbb76e7f6e8e4339628