unflake innodb_cleanup_dirty_tablespace
The problem with the code as is that in case of non-clean shutdown
innodb's procedure to restart the dictionary is slightly different
causing the fil_cleanup_dirty_tablespace-function never to be called.

This patch adds calls to fil_cleanup_dirty_tablespace also from
fil_space_for_table_exists_in_mem which is used in this scenario.

The reason for the flakiness was that the testcase did "shutdown 10"
which gives mysqld 10 seconds to shutdown, and then kill -9 it.
Due to the changes introduced by log_processlist_thread, it became
common for 10 seconds to no be enough.
(this problem is addressed in other commit)

Change-Id: I1363663c436c105033ee0adb9f2ac21ac78bf39c
Reviewed-by: Pavel Ivanov <pivanof@google.com>