make innochecksum work on compressed tables
this patch ports the work that facebook has performed
to make innochecksum handle compressed tables.
the basic idea is to use actual innodb-code to perform
checksum verification rather than duplicating in innochecksum.cc.
to make this work, innodb code has been annotated with
lots of #ifndef UNIV_INNOCHECKSUM so that it can be
compiled outside of storage/innobase.

a new testcase is also added that verifies that innochecksum
works on compressed/non-compressed tables.

the actual steps to produce this patch are:
- take innochecksum from 5.6.14
- apply changes in innodb from facebook patches needed to make innochecksum compile
- apply changes in innochecksum from facebook patches
- add handcrafted testcase

the referenced facebook patches used are:
https://github.com/facebook/mysql-5.6/commit/91e25120e75272db4cdbc07d0e45877d9dea5715
https://github.com/facebook/mysql-5.6/commit/847fe76ea5239b09fa361b023c56e6be76d32046
https://github.com/facebook/mysql-5.6/commit/1135628a5a9b3412621b93233478f3804bcef51a
https://github.com/facebook/mysql-5.6/commit/4dbf7c240ce2f08b7d6572d9452c9779ce90641c

Change-Id: I20c7bc7e26094f5bdb7346d905084abd7ad2f4d8