Expose total per-table space file size as Max_data_length
Although SHOW TABLE STATUS exposes, for InnoDB tables, the size of the
clustered and secondary indexes and the amount of free space, it does
not expose, nor it is possible to calculate, the total size of the
tablespace to which the table belongs, because any internal overhead is
deducted from the exposed sizes.

This change repurposes the Max_data_length field of SHOW TABLE STATUS to
expose the actual total size of the tablespace to which an InnoDB table
belongs to. Previously, this field was left unused (zeroed) for such
tables.

Additionally several tests which included the output of max_data_length
have been cleaned up (since the value wasn't being meaningfully tested
before anyway):

  innodb_stats_create_on_corrupted
  innodb_stats_fetch
  innodb_stats_fetch_corrupted
  innodb_stats_fetch_nonexistent

This change is ported from Twitter MySQL 5.5:

  https://github.com/twitter/mysql/commit/6baacfe

Change-Id: I606c6004e5dc97321cc46bb997e88ad690913a73