Always physically close sockets when connections are killed.
Sometimes after execution of KILL command either SIGUSR1 to the thread could be
lost or the killed flag on THD could be reset somewhere which allowed the killed
connection to continue executing statements. Instead of chasing places where
kill signal can be lost we'll just always close socket on the killed
connection.

Change-Id: I34f6505c3174af00eff58ffc9cd28621f7d602c0
Reviewed-by: Jeremy Cole <jeremycole@google.com>