Ensure time zone data is correctly generated
Ensure time zone data generated does not have any overly long
abbreviations. The time zone abbreviation may contain
'Local time zone must be set--see zic manual page'
which is longer than the CHAR(8) that MySQL allocates for it.
This causes a failure on INSERT of this record, which causes
a total failure of the script when running in strict mode.
In the interest of a simpler change, and given that the problematic
data is a human-directed message rather than a valid abbreviation,
change any such overly-long abbreviations to 'UNKNOWN' when the
SQL is generated.

Also remove the ALTER TABLE ... ORDER BY statements from the time zone SQL
generated from sql/tztime.c, as they are incompatible with InnoDB.

Change-Id: I9ff0e6c192ec113d65bee9294c2e70723195eaeb