Add replication-safe version of ON DUPLICATE KEY UPDATE
Specifying the key name in ON DUPLICATE KEY UPDATE resolves the
statement replication safety issues by allowing the ON DUPLICATE
behavior to apply to exactly one unique key. Add a new syntax of
the form:

INSERT ... ON DUPLICATE KEY (keyname) UPDATE ...

Statements supplying a key name are not marked as unsafe for
statement-based replication.

Change-Id: I5500b096966c8f588cd8b76ac20312a2aed9d33f