Add possibility to suppress replication safety warning.
Add a special SUPPRESS_SAFETY_WARNING clause to INSERT, UPDATE, DELETE and
CREATE TABLE statements which will suppress certain types of warnings about
statements being replication unsafe. This is intended to be used by users
when they are sure that the statement is actually safe, but due to a coarse
nature of safety detection in MySQL/MariaDB it's still marked as unsafe.

Also a flag --permit_safety_suppression is added to be able to prohibit usage
of SUPPRESS_SAFETY_WARNING clause by users abusing it and not listening to
request to stop using it without consultation.

The effect of SUPPRESS_SAFETY_WARNING clause is also coarse and so it can
suppress the replication safety warning in some cases in statements that
cannot ever be safe. But that's a trade-off that we'll take to not try
implementing the real correct solution. Real solution is really complicated.
MySQL has related bug (http://bugs.mysql.com/bug.php?id=42415) open for 5 years
now and even some patches were attempted to be done, but they were rejected
due to being wrong in some corner cases.

Change-Id: Id3a3e57511b9a17e9894f03a91df134fb609de7c