Comment out part of rpl_stop_slave test that is racy.
This part of the test also seems to be out of sync with the comment for it,
or there is a bug causing it to not actually be testing what it claims. You
can see this by adding SHOW BINLOG EVENTS and observing that the
non-transactional table insert is put into its own transaction. This causes
the slave sql thread to safely rollback the transaction containing only the
transactional table inserts (not wait for the commit as the comment claims).

Change-Id: Iadf452f78b1335891e4c742d03c88ca19c466498