Set the length of a field name in an Item
If the name_length is not set, name will store the field name, but since
name_length is 0, it will not be usable by other subsystems, such as the
Ed_* (execute direct) interface for internally executing queries. This
will result in empty field names being propagated in the results of such
internal queries.

Change-Id: I7697578cd5344d167869c102c8c1cd2f539a5f70