Change type of information_schema.processlist.info column to longblob.
When info column is longtext MySQL tries to convert it from table's character
set into connection's character set. information_schema.processlist has utf8
as its character set, but the query string that MySQL tries to convert comes
from connections that don't necessarily have utf8 as there current character
set, so conversion may not be possible. Even worse conversion from each
connection's character set into utf8 may not be possible, because if current
connection's character set is the same as character set of tables it works with
then MySQL doesn't check for validity of the incoming query strings. So those
may be not well formed. Note also that it's not enough to always have utf8 as
current connection's character set when you read from processlist table, because
even though no conversion from utf8 to utf8 is necessary MySQL still checks
that the query string is a well formed utf8, which it apparently is often not.

Change-Id: I525ab9d3c22d75e5953f3462c4d22b9abe33aced